ProductManagement/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyMySQL==1.1.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
django-filter==23.3
drf-spectacular==0.26.5
python-consul==1.1.0